requirements.txt
streamlit==1.28.2
PyPDF2==3.0.1
transformers==4.35.2
torch==2.1.1
google-generativeai==0.3.1